decimal separator . files changed:  M Makefile.sources  M src/main/java/managers/MoneyManager.java MM src/main/java/managers/MongoManager.java  M src/main/java/managers/TrelloManager.java AM src/main/java/managers/mongomanager/MongoManagerHelper.java  M src/main/java/util/Util.java  M trello.json
trello card: https://trello.com/c/iWI3D4Dn

+++ src/main/java/managers/TrelloManager.java
@@ -192,7 +192,7 @@ public class TrelloManager extends AbstractManager{
 	public String removecards(String rem) throws Exception {
 		String[] split = rem.split(" ",2);
 		String listId = ta_.findListByName(HABITBOARDID, "TODO");
-		int count = Util.SimpleEval(split[0]);
+		int count = (int) Util.SimpleEval(split[0]);
 		String name = split[1];
 		JSONArray array = ta_.getCardsInList(listId);
